set actuators typo to np.float32 to avoid casting errors
(cherry picked from commit 4e64728f1ab03733410101d95055fbc0cdb8d2ff)

+++ racecar_gym/bullet/actuators.py
@@ -59,7 +59,7 @@ class Motor(BulletActuator[Tuple[float, float]]):
             )
 
     def space(self) -> gym.Space:
-        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float64)
+        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float32)
 
 
 class Speed(BulletActuator[Tuple[float, float]]):
@@ -90,7 +90,7 @@ class Speed(BulletActuator[Tuple[float, float]]):
             )
 
     def space(self) -> gym.Space:
-        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float64)
+        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float32)
 
 
 class SteeringWheel(BulletActuator[float]):
@@ -114,4 +114,4 @@ class SteeringWheel(BulletActuator[float]):
             )
 
     def space(self) -> gym.Space:
-        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float64)
+        return gym.spaces.Box(low=-1.0, high=1.0, shape=(1,), dtype=np.float32)
